Compile VIM with Python 2 support only
Compiling with support for both allows only the one (Python 2 or 3)
loaded first to run.  YouCompleteMe seems to load Python 3 if it is
available, so clang-format, which requires Python 2, doesn't work.
However, YouCompleteMe works with Python 2, so compiling with support
only for Python 2 forces it to use that such that clang-format is able
to use Python 2.